remove duplicated functions. fix unit tests
