+++ lambdas/check_new_areas/lambda_function.py
@@ -7,6 +7,7 @@ from typing import Any, Dict, Iterator, List, Optional, Tuple
 
 import boto3
 import requests
+import yaml
 from requests import Response
 from shapely.wkb import dumps
 from shapely.geometry import shape, Polygon
@@ -26,19 +27,11 @@ else:
 S3_CLIENT = boto3.client("s3")
 TOKEN: str = get_token()
 PENDING_AOI_NAME = "pending_user_areas"
-PENDING_AOI_ANALYSES = {
-    "dev": {
-        "annualupdate_minimal": {
-            "change": "206938be-12d9-47b7-9865-44244bfb64d6",
-            "summary": "0eead72d-1ad7-4c0f-93c4-793a07cd2e3d",
-        },
-        "gladalerts": {
-            "daily_alerts": "722d90b2-e989-48ca-ba27-f7a8e236ea44",
-            "weekly_alerts": "153a2ba7-cff6-4b06-bd76-279271c4ddea",
-            "summary": "28e44bd2-cd13-4587-9259-1ba235a82d28",
-        },
-    }
-}
+
+DIRNAME = os.path.dirname(__file__)
+
+with open(os.path.join(DIRNAME, "analysis_config.yaml"), "r") as config:
+    PENDING_AOI_ANALYSES = yaml.safe_load(config)
 
 
 def handler(
