+++ lambdas/check_new_areas/requirements.txt
@@ -1,2 +1,3 @@
 requests~=2.22.0
 shapely~=1.6.4.post2
+PyYAML~=5.2
