+++ setup.py
@@ -7,5 +7,10 @@ setup(
     packages=["geotrellis_summary_update", "lambdas"],
     author="Justin Terry",
     license="MIT",
-    install_requires=["boto3~=1.10.7", "requests~=2.22.0", "shapely~=1.6.4.post2"],
+    install_requires=[
+        "boto3~=1.10.7",
+        "PyYAML~=5.2",
+        "requests~=2.22.0",
+        "shapely~=1.6.4.post2",
+    ],
 )
